% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func.abun.R
\name{calc.abun}
\alias{calc.abun}
\title{Cálcular abundancias, actividades estratos y diversidad}
\usage{
calc.abun(PC_lista)
}
\arguments{
\item{PC_lista}{lista con dataframes de distintos proyectos de una hoja (ej: PCTS_PEEM_OAX).}
}
\value{
Archivos xlsx con los resultados por proyecto
}
\description{
Aplica calculos de abundancia relativa de cada especie, actividades, diversidad y estratos.
}
\author{
Biol. Michell Romero - Coordinador del área de Monitoreo de aves TCT 🦉
}
