% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_CV.R
\name{calc.cv}
\alias{calc.cv}
\title{Cálcular resultados del método "Conductas de vuelo"}
\usage{
calc.cv(CV_lista)
}
\arguments{
\item{CV_lista}{lista con dataframes de distintos proyectos de una hoja (ej: CV_PEEM; CV_OAX; CV_BH).}
}
\value{
Archivos xlsx con los resultados por proyecto.
}
\description{
Aplica calculos de vuelos por categoría de riesgo, número de registros por tipo de vuelo así como por dirección de vuelo.
}
\author{
Biol. Michell Romero - Coordinador del área de Monitoreo de aves TCT 🦉
}
