% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extraerspp.R
\name{extraer_spp}
\alias{extraer_spp}
\title{Funci贸n para extraer informaci贸n de spp de un listado potencial}
\usage{
extraer_spp(lista_hojas, inventario)
}
\arguments{
\item{lista_hojas}{lista con dataframes de donde se obtendr谩n las spp de consulta.}

\item{inventario}{Inventario limpio (clean_names()) con informaci贸n de las spp objetivo y sus variables en columnas (p.ej. NOM059, IUCN, CITES).}
}
\value{
Las especies consultadas con toda la informaci贸n del inventario (columnas).
}
\description{
Extraer las especies 煤nicas de cada hoja o df y las usa para filtrar la informaci贸n de un inventario general.
}
\author{
Biol. Michell Romero - Coordinador del 谩rea de Monitoreo de aves TCT 
}
