% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filtrado.R
\name{filtrado}
\alias{filtrado}
\title{Filtrado de datos por proyecto y periodos}
\usage{
filtrado(df, proyectos_filtrar, anual, mes)
}
\arguments{
\item{df}{Dataframe combinado o único (ej: CV_combinadas, PCTS_combinado).}

\item{proyectos_filtrar}{Vector de nombres de proyectos a incluir (ej: c("PEEM_PE", "PEEM_LT", "OAX_I", "BH")).}

\item{anual}{Año de la fecha a filtrar.}

\item{mes}{Mes de la fecha a filtrar (número).}
}
\value{
Un dataframe, dividido por la columna 'Proyecto'.
}
\description{
Aplica filtrados de datos por proyecto y una fecha en específico
}
\author{
Biol. Michell Ivann Romero Pacheco
}
