% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codificacionPC.R
\name{tipo_dato.PCTS}
\alias{tipo_dato.PCTS}
\title{Asignación del tipo de dato a cada una de las variables de la hoja PC_TS}
\usage{
tipo_dato.PCTS(lista_archivos)
}
\arguments{
\item{lista_archivos}{lista con dataframes de distintos proyectos de una hoja (ej: PCTS_PEEM_OAX).}
}
\value{
El mismo dataframe pero con el tipo de dato correcto.
}
\description{
Aplica a cada columna el tipo de dato que le corresponde, integer, string, etc.
}
\author{
Biol. Michell Romero - Coordinador del área de Monitoreo de aves TCT 🦉
}
